refactor(step2): package 구성 변경
