+++ src/main/kotlin/MineSweeperApplication.kt
@@ -1,6 +1,6 @@
-import domain.MineFieldGenerator
+import domain.field.MineFieldGenerator
 import domain.RandomMinePositionsSelector
-import domain.Rectangle
+import domain.field.Rectangle
 import view.getHeightOfField
 import view.getMinesCount
 import view.getWidthOfField
@@ -9,7 +9,8 @@ import view.printMineField
 
 fun main() {
     val rectangle = Rectangle(getWidthOfField(), getHeightOfField())
-    val mineField = MineFieldGenerator(RandomMinePositionsSelector()).create(rectangle, getMinesCount())
+    val mineField = MineFieldGenerator(RandomMinePositionsSelector())
+        .create(rectangle, getMinesCount())
     printGameStart()
     printMineField(mineField)
 }
