+++ src/main/kotlin/domain/MinePositionsSelectStrategy.kt
@@ -1,5 +1,7 @@
 package domain
 
+import domain.block.Position
+
 interface MinePositionsSelectStrategy {
 
     fun getMinePositionsFrom(positions: List<Position>, count: Int): List<Position>
