+++ src/main/kotlin/domain/block/NormalBlock.kt
@@ -1,10 +1,14 @@
-package domain
+package domain.block
 
 data class NormalBlock(
     override val position: Position,
     val mineCount: Int
 ) : Block {
-    constructor(xPosition: Int, yPosition: Int, mineCount: Int) : this(Position.of(xPosition, yPosition), mineCount)
+    constructor(xPosition: Int, yPosition: Int, mineCount: Int) : this(
+        Position.of(
+            xPosition,
+            yPosition
+        ), mineCount)
 
     override fun isMine(): Boolean = false
