+++ src/main/kotlin/domain/field/MineField.kt
@@ -1,4 +1,6 @@
-package domain
+package domain.field
+
+import domain.block.Block
 
 data class MineField(
     private val rectangle: Rectangle,
