+++ src/test/kotlin/domain/RandomMinePositionsSelectorTest.kt
@@ -1,5 +1,6 @@
 package domain
 
+import domain.block.Position
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.jupiter.api.Test
